requirements.txt
streamlit
pandas
feedparser
requests
python-dateutil
google-generativeai
beautifulsoup4
